+++ twitter4j-v2-support/src/main/kotlin/twitter4j/ReplySettings.kt
@@ -1,0 +1,6 @@
+package twitter4j
+
+enum class ReplySettings(val value: String) {
+    MentionedUsers("mentionedUsers"),
+    Following("following")
+}